Allow to print fmf identifier in tmt tests show
Part of `tmt tests show --verbose`
