+++ tmt/base.py
@@ -6,6 +6,7 @@ import os
 import re
 import fmf
 import pprint
+import subprocess
 
 import tmt.steps
 import tmt.utils
@@ -63,6 +64,11 @@ class Node(tmt.utils.Common):
             raise tmt.utils.SpecificationError(
                 f"Invalid '{key}' in {class_name} '{self.name}' (should be "
                 f"a '{expected.__name__}', got a '{type(value).__name__}').")
+
+    def _fmf_id(self):
+        """ Show fmf identifier """
+        echo(tmt.utils.format(
+            'fmf-id', self.fmf_id, key_color='magenta'))
 
     @classmethod
     def _save_context(cls, context):
@@ -225,6 +231,8 @@ class Test(Node):
                 echo(tmt.utils.format(key, value))
         if self.opt('verbose'):
             self._sources()
+            self._fmf_id()
+
 
     def lint(self):
         """ Check test against the L1 metadata specification. """
@@ -272,6 +280,34 @@ class Test(Node):
         # Common node export otherwise
         else:
             return super(Test, self).export(format_, keys)
+
+    @property
+    def fmf_id(self):
+        """ Valid fmf identifier for test case """
+
+        def run(command):
+            """ Run command, return output """
+            result = subprocess.run(command.split(), capture_output=True)
+            return result.stdout.strip().decode("utf-8")
+
+        fmf_id = {'name': self.name}
+
+        # Prepare url (for now handle just the most common schemas)
+        origin = run('git config --get remote.origin.url')
+        fmf_id['url'] = tmt.utils.public_git_url(origin)
+
+        # Get the ref (skip for master as it is the default)
+        ref = run('git rev-parse --abbrev-ref HEAD')
+        if ref != 'master':
+            fmf_id['ref'] = ref
+
+        # Construct path (if different from git root)
+        git_root = run('git rev-parse --show-toplevel')
+        fmf_root = self.node.root
+        if git_root != fmf_root:
+            fmf_id['path'] = os.path.join('/', os.path.relpath(fmf_root, git_root))
+
+        return fmf_id
 
 
 class Plan(Node):
